+++ index.html
@@ -12,11 +12,12 @@
   <link rel="stylesheet" href="./styles/animation.css">
   <link rel="stylesheet" href="./styles/dialog.css">
   <script type="module" src="./src/main.js"></script>
+  <script type="module" src="./src/hero.js"></script>
 </head>
 
 <body>
   <header id="header">
-    <h1><a href="./index.html">PokéSafari</a> <button class="mobile-button">☰</button></h1>
+    <h1><a href="./index.html">PokéSafari</a> <span><span id="player"></span><button class="mobile-button">☰</button></span></h1>
     <hr>
   </header>
   <nav>
