+++ pokedex.html
@@ -11,12 +11,13 @@
   <link rel="stylesheet" href="./styles/footer.css">
   <link rel="stylesheet" href="./styles/animation.css">
   <link rel="stylesheet" href="./styles/pokedex.css">
+  <script type="module" src="./src/main.js"></script>
   <script type="module" src="./src/pokedex.js"></script>
 </head>
 
 <body>
   <header id="header">
-    <h1><a href="./index.html">PokéSafari</a> <button class="mobile-button">☰</button></h1>
+    <h1><a href="./index.html">PokéSafari</a> <span><span id="player"></span><button class="mobile-button">☰</button></span></h1>
     <hr>
   </header>
   <nav>
