+++ styles/nav.css
@@ -64,6 +64,12 @@ nav {
     z-index: 10;
 }
 
+#player{
+    font-size: 1rem;
+    font-weight: 400;
+    vertical-align: middle;
+}
+
 @media screen and (min-width: 612px){
 
     header {
