+++ team.html
@@ -15,7 +15,7 @@
 
 <body>
   <header>
-    <h1><a href="./index.html">PokéSafari</a> <button class="mobile-button">☰</button></h1>
+    <h1><a href="./index.html">PokéSafari</a> <span><span id="player"></span><button class="mobile-button">☰</button></span></h1>
     <hr>
   </header>
   <nav>
